major reformation of generated class structure
