+++ tests/DataSource/Course/CourseRow.php
@@ -1,21 +1,8 @@
 <?php
-/**
- * This file was generated by Atlas. Changes will be overwritten.
- */
 declare(strict_types=1);
 
 namespace Atlas\Table\DataSource\Course;
 
-use Atlas\Table\Row;
-
-/**
- * @property mixed $course_subject CHAR(4)
- * @property mixed $course_number INT
- * @property mixed $title VARCHAR(20)
- */
-class CourseRow extends Row
+class CourseRow extends _\CourseRow_
 {
-    protected mixed $course_subject = null;
-    protected mixed $course_number = null;
-    protected mixed $title = null;
 }
