+++ tests/DataSource/Course/CourseTableSelect.php
@@ -3,12 +3,6 @@ declare(strict_types=1);
 
 namespace Atlas\Table\DataSource\Course;
 
-use Atlas\Table\TableSelect;
-
-/**
- * @method CourseRow|null fetchRow()
- * @method CourseRow[] fetchRows()
- */
-class CourseTableSelect extends TableSelect
+class CourseTableSelect extends _\CourseTableSelect_
 {
 }
