+++ tests/DataSource/Course/_/CourseRow_.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * This file was generated by Atlas. Changes will be overwritten.
+ */
+declare(strict_types=1);
+
+namespace Atlas\Table\DataSource\Course\_;
+
+use Atlas\Table\Row;
+
+/**
+ * @property mixed $course_subject CHAR(4)
+ * @property mixed $course_number INT
+ * @property mixed $title VARCHAR(20)
+ */
+abstract class CourseRow_ extends Row
+{
+    protected mixed $course_subject = null;
+    protected mixed $course_number = null;
+    protected mixed $title = null;
+}
