+++ tests/DataSource/Course/_/CourseTableEvents_.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * This file was generated by Atlas. Changes will be overwritten.
+ */
+declare(strict_types=1);
+
+namespace Atlas\Table\DataSource\Course\_;
+
+use Atlas\Query\Delete;
+use Atlas\Query\Insert;
+use Atlas\Query\Select;
+use Atlas\Query\Update;
+use Atlas\Table\Row;
+use Atlas\Table\Table;
+use Atlas\Table\TableEvents;
+use PDOStatement;
+
+abstract class CourseTableEvents_ extends TableEvents
+{
+}
