+++ tests/DataSource/Course/_/CourseTableSelect_.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * This file was generated by Atlas. Changes will be overwritten.
+ */
+declare(strict_types=1);
+
+namespace Atlas\Table\DataSource\Course\_;
+
+use Atlas\Table\TableSelect;
+
+/**
+ * @method CourseRow|null fetchRow()
+ * @method CourseRow[] fetchRows()
+ */
+class CourseTableSelect_ extends TableSelect
+{
+}
