+++ tests/DataSource/Employee/EmployeeRow.php
@@ -1,21 +1,8 @@
 <?php
-/**
- * This file was generated by Atlas. Changes will be overwritten.
- */
+declare(strict_types=1);
+
 namespace Atlas\Table\DataSource\Employee;
 
-use Atlas\Table\Row;
-
-/**
- * @property mixed $id INTEGER
- * @property mixed $name VARCHAR(10) NOT NULL
- * @property mixed $building INTEGER
- * @property mixed $floor INTEGER
- */
-class EmployeeRow extends Row
+class EmployeeRow extends _\EmployeeRow_
 {
-    protected mixed $id = null;
-    protected mixed $name = null;
-    protected mixed $building = null;
-    protected mixed $floor = null;
 }
