+++ tests/DataSource/Employee/EmployeeTable.php
@@ -1,92 +1,8 @@
 <?php
-/**
- * This file was generated by Atlas. Changes will be overwritten.
- */
 declare(strict_types=1);
 
 namespace Atlas\Table\DataSource\Employee;
 
-use Atlas\Table\Table;
-
-/**
- * @method EmployeeRow|null fetchRow($primaryVal)
- * @method EmployeeRow[] fetchRows(array $primaryVals)
- * @method EmployeeTableSelect select(array $whereEquals = [])
- * @method EmployeeRow newRow(array $cols = [])
- * @method EmployeeRow newSelectedRow(array $cols)
- */
-class EmployeeTable extends Table
+class EmployeeTable extends _\EmployeeTable_
 {
-    const DRIVER = 'sqlite';
-
-    const NAME = 'employees';
-
-    const COLUMNS = [
-        'id' => [
-            'name' => 'id',
-            'type' => 'INTEGER',
-            'size' => null,
-            'scale' => null,
-            'notnull' => false,
-            'default' => null,
-            'autoinc' => true,
-            'primary' => true,
-            'options' => null,
-        ],
-        'name' => [
-            'name' => 'name',
-            'type' => 'VARCHAR',
-            'size' => 10,
-            'scale' => null,
-            'notnull' => true,
-            'default' => null,
-            'autoinc' => false,
-            'primary' => false,
-            'options' => null,
-        ],
-        'building' => [
-            'name' => 'building',
-            'type' => 'INTEGER',
-            'size' => null,
-            'scale' => null,
-            'notnull' => false,
-            'default' => null,
-            'autoinc' => false,
-            'primary' => false,
-            'options' => null,
-        ],
-        'floor' => [
-            'name' => 'floor',
-            'type' => 'INTEGER',
-            'size' => null,
-            'scale' => null,
-            'notnull' => false,
-            'default' => null,
-            'autoinc' => false,
-            'primary' => false,
-            'options' => null,
-        ],
-    ];
-
-    const COLUMN_NAMES = [
-        'id',
-        'name',
-        'building',
-        'floor',
-    ];
-
-    const COLUMN_DEFAULTS = [
-        'id' => null,
-        'name' => null,
-        'building' => null,
-        'floor' => null,
-    ];
-
-    const PRIMARY_KEY = [
-        'id',
-    ];
-
-    const AUTOINC_COLUMN = 'id';
-
-    const AUTOINC_SEQUENCE = null;
 }
