+++ tests/DataSource/Employee/EmployeeTableSelect.php
@@ -3,12 +3,6 @@ declare(strict_types=1);
 
 namespace Atlas\Table\DataSource\Employee;
 
-use Atlas\Table\TableSelect;
-
-/**
- * @method EmployeeRow|null fetchRow()
- * @method EmployeeRow[] fetchRows()
- */
-class EmployeeTableSelect extends TableSelect
+class EmployeeTableSelect extends _\EmployeeTableSelect_
 {
 }
