+++ tests/DataSource/Employee/_/EmployeeRow_.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * This file was generated by Atlas. Changes will be overwritten.
+ */
+declare(strict_types=1);
+
+namespace Atlas\Table\DataSource\Employee\_;
+
+use Atlas\Table\Row;
+
+/**
+ * @property mixed $id INTEGER
+ * @property mixed $name VARCHAR(10) NOT NULL
+ * @property mixed $building INTEGER
+ * @property mixed $floor INTEGER
+ */
+abstract class EmployeeRow_ extends Row
+{
+    protected mixed $id = null;
+    protected mixed $name = null;
+    protected mixed $building = null;
+    protected mixed $floor = null;
+}
