+++ tests/DataSource/Employee/_/EmployeeTableSelect_.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * This file was generated by Atlas. Changes will be overwritten.
+ */
+declare(strict_types=1);
+
+namespace Atlas\Table\DataSource\Employee\_;
+
+use Atlas\Table\TableSelect;
+
+/**
+ * @method EmployeeRow|null fetchRow()
+ * @method EmployeeRow[] fetchRows()
+ */
+abstract class EmployeeTableSelect_ extends TableSelect
+{
+}
