+++ tests/DataSource/Nopkey/NopkeyRow.php
@@ -1,19 +1,8 @@
 <?php
-/**
- * This file was generated by Atlas. Changes will be overwritten.
- */
 declare(strict_types=1);
 
 namespace Atlas\Table\DataSource\Nopkey;
 
-use Atlas\Table\Row;
-
-/**
- * @property mixed $name VARCHAR(255)
- * @property mixed $email VARCHAR(255)
- */
-class NopkeyRow extends Row
+class NopkeyRow extends _\NopkeyRow_
 {
-    protected mixed $name;
-    protected mixed $email;
 }
