+++ tests/DataSource/Nopkey/NopkeyTableEvents.php
@@ -3,15 +3,6 @@ declare(strict_types=1);
 
 namespace Atlas\Table\DataSource\Nopkey;
 
-use Atlas\Query\Delete;
-use Atlas\Query\Insert;
-use Atlas\Query\Select;
-use Atlas\Query\Update;
-use Atlas\Table\Row;
-use Atlas\Table\Table;
-use Atlas\Table\TableEvents;
-use PDOStatement;
-
-class NopkeyTableEvents extends TableEvents
+class NopkeyTableEvents extends _\NopkeyTableEvents_
 {
 }
