+++ tests/DataSource/Nopkey/NopkeyTableSelect.php
@@ -3,12 +3,6 @@ declare(strict_types=1);
 
 namespace Atlas\Table\DataSource\Nopkey;
 
-use Atlas\Table\TableSelect;
-
-/**
- * @method NopkeyRow|null fetchRow()
- * @method NopkeyRow[] fetchRows()
- */
-class NopkeyTableSelect extends TableSelect
+class NopkeyTableSelect extends _\NopkeyTableSelect_
 {
 }
