+++ tests/DataSource/Nopkey/_/NopkeyRow_.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * This file was generated by Atlas. Changes will be overwritten.
+ */
+declare(strict_types=1);
+
+namespace Atlas\Table\DataSource\Nopkey\_;
+
+use Atlas\Table\Row;
+
+/**
+ * @property mixed $name VARCHAR(255)
+ * @property mixed $email VARCHAR(255)
+ */
+abstract class NopkeyRow_ extends Row
+{
+    protected mixed $name;
+    protected mixed $email;
+}
