+++ tests/DataSource/Nopkey/_/NopkeyTableEvents_.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * This file was generated by Atlas. Changes will be overwritten.
+ */
+declare(strict_types=1);
+
+namespace Atlas\Table\DataSource\Nopkey\_;
+
+use Atlas\Table\TableEvents;
+
+abstract class NopkeyTableEvents_ extends TableEvents
+{
+}
