+++ tests/DataSource/Nopkey/_/NopkeyTableSelect_.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * This file was generated by Atlas. Changes will be overwritten.
+ */
+declare(strict_types=1);
+
+namespace Atlas\Table\DataSource\Nopkey\_;
+
+use Atlas\Table\TableSelect;
+
+/**
+ * @method NopkeyRow|null fetchRow()
+ * @method NopkeyRow[] fetchRows()
+ */
+abstract class NopkeyTableSelect_ extends TableSelect
+{
+}
